+++ src/bulirsch/cel.rs
@@ -7,8 +7,7 @@ use num_traits::Float;
 
 use super::BulirschConst;
 
-// Reference: Bulirsch, “Numerical Calculation of Elliptic Integrals and Elliptic Functions III.”
-/// Compute [complete elliptic integral in Bulirsch form](https://dlmf.nist.gov/19.2#iii).
+/// Computes [complete elliptic integral in Bulirsch form](https://dlmf.nist.gov/19.2#iii).
 /// ```text
 ///                       π/2                                                   
 ///                      ⌠            a cos²(ϑ) + b sin²(ϑ)               
@@ -16,9 +15,44 @@ use super::BulirschConst;
 ///                      ⎮                         ______________________
 ///                      ⌡ (cos²(ϑ) + p sin²(ϑ)) ╲╱ cos²(ϑ) + kc² sin²(ϑ)    
 ///                     0                                                   
-/// where kc ≠ 0, p ≠ 0
 /// ```
-/// Note that kc² = mc = 1 - m. The precision can be adjusted by overwriting the trait [super::BulirschConst].
+///
+/// ## Parameters
+/// - kc: complementary modulus. kc ∈ ℝ, kc ≠ 0.
+/// - p ∈ ℝ, p ≠ 0
+/// - a ∈ ℝ
+/// - b ∈ ℝ
+///
+/// The precision of the function can be adjusted by overwriting the trait [super::BulirschConst].
+/// The default is set according to the original literature by [Bulirsch](https://doi.org/10.1007/BF02165405) for [f64] and [f32].
+///
+/// ## Domain
+/// - Returns error if kc = 0 or p = 0.
+/// - Returns the principal value for p < 0.
+///
+/// ## Graph
+/// ![General Complete Elliptic Integral](https://github.com/p-sira/ellip/blob/main/figures/cel_plot.svg?raw=true)
+///
+/// [Interactive Plot](https://github.com/p-sira/ellip/blob/main/figures/cel_plot.html)
+///
+/// # Related Functions
+/// With kc² = 1 - m and p = 1 - n,
+/// - [ellipk](crate::ellipk)(m) = [cel](crate::cel)(kc, 1, 1, 1) = [cel1](crate::cel1)(kc)
+/// - [ellipe](crate::ellipe)(m) = [cel](crate::cel)(kc, 1, 1, kc²) = [cel2](crate::cel2)(kc, 1, kc²)
+/// - [ellipd](crate::ellipd)(m) = [cel](crate::cel)(kc, 1, 0, 1)
+/// - [ellippi](crate::ellippi)(n, m) = [cel](crate::cel)(kc, p, 1, 1)
+///
+/// # Examples
+/// ```
+/// use ellip::{cel, util::assert_close};
+///
+/// assert_close(cel(0.5, 1.0, 1.0, 1.0).unwrap(), 2.1565156474996434, 1e-15);
+/// ```
+///
+/// # References
+/// - Bulirsch, R. “Numerical Calculation of Elliptic Integrals and Elliptic Functions. III.” Numerische Mathematik 13, no. 4 (August 1, 1969): 305–15. <https://doi.org/10.1007/BF02165405>.
+/// - Carlson, B. C. “DLMF: Chapter 19 Elliptic Integrals.” Accessed February 19, 2025. <https://dlmf.nist.gov/19>.
+///
 pub fn cel<T: Float + BulirschConst>(kc: T, p: T, a: T, b: T) -> Result<T, &'static str> {
     if kc == zero!() {
         return Err("cel: kc cannot be zero.");
@@ -71,8 +105,7 @@ pub fn cel<T: Float + BulirschConst>(kc: T, p: T, a: T, b: T) -> Result<T, &'sta
     Ok(pi_2!() * (a * m + b) / (m * (m + p)))
 }
 
-// Reference: Bulirsch, “Numerical Calculation of Elliptic Integrals and Elliptic Functions.”
-/// Compute [complete elliptic integral of the first kind in Bulirsch's form](https://link.springer.com/article/10.1007/bf01397975).
+/// Computes [complete elliptic integral of the first kind in Bulirsch's form](https://link.springer.com/article/10.1007/bf01397975).
 /// ```text
 ///               π/2                                                   
 ///              ⌠               dϑ              
@@ -80,10 +113,37 @@ pub fn cel<T: Float + BulirschConst>(kc: T, p: T, a: T, b: T) -> Result<T, &'sta
 ///              ⎮      ______________________
 ///              ⌡   ╲╱ cos²(ϑ) + kc² sin²(ϑ)    
 ///             0                                                   
-/// where kc ≠ 0
 /// ```
 ///
-/// Note that kc² = mc = 1 - m. The precision can be adjusted by overwriting the trait [super::BulirschConst].
+/// ## Parameters
+/// - kc: complementary modulus. kc ∈ ℝ, kc ≠ 0.
+///
+/// The precision of the function can be adjusted by overwriting the trait [super::BulirschConst].
+/// The default is set according to the original literature by [Bulirsch](https://doi.org/10.1007/BF02165405) for [f64] and [f32].
+///
+/// ## Domain
+/// - Returns error if kc = 0.
+///
+/// ## Graph
+/// ![Bulirsch's Complete Elliptic Integral of the First Kind](https://github.com/p-sira/ellip/blob/main/figures/cel1_plot.svg?raw=true)
+///
+/// [Interactive Plot](https://github.com/p-sira/ellip/blob/main/figures/cel1_plot.html)
+///
+/// # Related Functions
+/// With kc² = 1 - m,
+/// - [ellipk](crate::ellipk)(m) = [cel](crate::cel)(kc, 1, 1, 1) = [cel1](crate::cel1)(kc)
+///
+/// # Examples
+/// ```
+/// use ellip::{cel1, util::assert_close};
+///
+/// assert_close(cel1(0.5).unwrap(), 2.1565156474996434, 1e-15);
+/// ```
+///
+/// # References
+/// - Bulirsch, Roland. “Numerical Calculation of Elliptic Integrals and Elliptic Functions.” Numerische Mathematik 7, no. 1 (February 1, 1965): 78–90. <https://doi.org/10.1007/BF01397975>.
+/// - Carlson, B. C. “DLMF: Chapter 19 Elliptic Integrals.” Accessed February 19, 2025. <https://dlmf.nist.gov/19>.
+///
 pub fn cel1<T: Float + BulirschConst>(kc: T) -> Result<T, &'static str> {
     if kc == zero!() {
         return Err("cel1: kc cannot be zero.");
@@ -108,8 +168,7 @@ pub fn cel1<T: Float + BulirschConst>(kc: T) -> Result<T, &'static str> {
     Ok(pi!() / m)
 }
 
-// Reference: Bulirsch, “Numerical Calculation of Elliptic Integrals and Elliptic Functions.”
-/// Compute [complete elliptic integral of the second kind in Bulirsch's form](https://link.springer.com/article/10.1007/bf01397975).
+/// Computes [complete elliptic integral of the second kind in Bulirsch's form](https://link.springer.com/article/10.1007/bf01397975).
 /// ```text
 ///                     π/2                           
 ///                    ⌠              a + b tan²(ϑ)
@@ -120,7 +179,37 @@ pub fn cel1<T: Float + BulirschConst>(kc: T) -> Result<T, &'static str> {
 /// where kc ≠ 0
 /// ```
 ///
-/// Note that kc² = mc = 1 - m. The precision can be adjusted by overwriting the trait [super::BulirschConst].
+/// ## Parameters
+/// - kc: complementary modulus. kc ∈ ℝ, kc ≠ 0.
+/// - a ∈ ℝ
+/// - b ∈ ℝ
+///
+/// The precision of the function can be adjusted by overwriting the trait [super::BulirschConst].
+/// The default is set according to the original literature by [Bulirsch](https://doi.org/10.1007/BF02165405) for [f64] and [f32].
+///
+/// ## Domain
+/// - Returns error if kc = 0.
+///
+/// ## Graph
+/// ![Bulirsch's Complete Elliptic Integral of the Second Kind](https://github.com/p-sira/ellip/blob/main/figures/cel2_plot.svg?raw=true)
+///
+/// [Interactive Plot](https://github.com/p-sira/ellip/blob/main/figures/cel2_plot.html)
+///
+/// # Related Functions
+/// With kc² = 1 - m,
+/// - [ellipe](crate::ellipe)(m) = [cel](crate::cel)(kc, 1, 1, kc²) = [cel2](crate::cel2)(kc, 1, kc²)
+///
+/// # Examples
+/// ```
+/// use ellip::{cel2, util::assert_close};
+///
+/// assert_close(cel2(0.5, 1.0, 1.0).unwrap(), 2.1565156474996434, 1e-15);
+/// ```
+///
+/// # References
+/// - Bulirsch, Roland. “Numerical Calculation of Elliptic Integrals and Elliptic Functions.” Numerische Mathematik 7, no. 1 (February 1, 1965): 78–90. <https://doi.org/10.1007/BF01397975>.
+/// - Carlson, B. C. “DLMF: Chapter 19 Elliptic Integrals.” Accessed February 19, 2025. <https://dlmf.nist.gov/19>.
+///
 pub fn cel2<T: Float + BulirschConst>(kc: T, a: T, b: T) -> Result<T, &'static str> {
     if kc == zero!() {
         return Err("cel2: kc cannot be zero.");
@@ -164,14 +253,13 @@ mod tests {
     #[test]
     fn test_cel() {
         fn _test(kc: f64, p: f64) {
-            let k = (1.0 - kc * kc).sqrt();
-            let m = k * k;
+            let m = 1.0 - kc * kc;
             let ellipk = ellipk(m).unwrap();
             let ellipe = ellipe(m).unwrap();
 
             // cel precision is low for K cases
-            assert_close!(ellipk, cel(kc, 1.0, 1.0, 1.0).unwrap(), 5e-12);
-            assert_close!(ellipe, cel(kc, 1.0, 1.0, kc * kc).unwrap(), 1e-14);
+            assert_close!(ellipk, cel(kc, 1.0, 1.0, 1.0).unwrap(), 2e-12);
+            assert_close!(ellipe, cel(kc, 1.0, 1.0, kc * kc).unwrap(), 1e-15);
             assert_close!(
                 (ellipe - kc * kc * ellipk) / m,
                 cel(kc, 1.0, 1.0, 0.0).unwrap(),
@@ -183,16 +271,16 @@ mod tests {
             let ellippi = ellippi(n, m).unwrap();
 
             assert_close!(
-                (ellipk - ellipe) / (k * k),
+                (ellipk - ellipe) / m,
                 cel(kc, 1.0, 0.0, 1.0).unwrap(),
                 6e-12
             );
             // cel precision is very low for PI cases
-            assert_close!(ellippi, cel(kc, p, 1.0, 1.0).unwrap(), 1e-11);
+            assert_close!(ellippi, cel(kc, p, 1.0, 1.0).unwrap(), 3.5e-12);
             assert_close!(
                 (ellippi - ellipk) / (1.0 - p),
                 cel(kc, p, 0.0, 1.0).unwrap(),
-                1e-11
+                3.5e-12
             );
         }
 
@@ -217,9 +305,8 @@ mod tests {
     #[test]
     fn test_cel1() {
         fn test_kc(kc: f64) {
-            let k = (1.0 - kc * kc).sqrt();
-            let m = k * k;
-            assert_close!(ellipk(m).unwrap(), cel1(kc).unwrap(), 5e-12);
+            let m = 1.0 - kc * kc;
+            assert_close!(ellipk(m).unwrap(), cel1(kc).unwrap(), 2e-12);
         }
 
         let linsp_neg = linspace(-1.0, -1e-3, 100);
@@ -231,8 +318,7 @@ mod tests {
     #[test]
     fn test_cel2() {
         fn test_kc(kc: f64) {
-            let k = (1.0 - kc * kc).sqrt();
-            let m = k * k;
+            let m = 1.0 - kc * kc;
             assert_close!(ellipe(m).unwrap(), cel2(kc, 1.0, kc * kc).unwrap(), 7.7e-16);
         }